minor cleaning
